requirements.txt
Flask==3.0.3
llama-cpp-python==0.3.1
python-dotenv==1.0.1
torch==2.0.0
torchvision==0.15.1
torchaudio==2.0.0
--extra-index-url https://download.pytorch.org/whl/cu118
git+https://github.com/m-bain/whisperx.git
ctranslate2==4.4.0